Using uint8_t for the volume setting API.

+++ drivers/cs43l22.h
@@ -37,8 +37,8 @@ struct cs4x_dac {
 
 int cs4x_init(struct cs4x_dac *dac, struct i2c_bus *i2c, uint8_t adr, int rst);
 int cs4x_output(struct cs4x_dac *dac, unsigned int out);
-int cs4x_master_volume(struct cs4x_dac *dac, float vol);
-int cs4x_headphone_volume(struct cs4x_dac *dac, float vol);
+int cs4x_master_volume(struct cs4x_dac *dac, uint8_t vol);
+int cs4x_headphone_volume(struct cs4x_dac *dac, uint8_t vol);
 int cs4x_speaker_volume(struct cs4x_dac *dac, uint8_t vol);
 
 //-----------------------------------------------------------------------------
